Add implementations for Rails 3.2 and 4.0.
